Add# DNSChecker

+++ checker.go
@@ -2,5 +2,5 @@ package main
 
 type Checker interface {
 	//New(Parameter, Expect interface{}) Checker
-	Check() (bool, uint, string)
+	Check() (bool, int64, string)
 }
